# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Soroban VRF API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type VrfRequestStatus =
  (typeof VrfRequestStatus)[keyof typeof VrfRequestStatus];

export const VrfRequestStatus = {
  pending: "pending",
  fulfilled: "fulfilled",
  failed: "failed",
} as const;

export interface VrfRequest {
  id: number;
  /** The input seed provided by the requester */
  alphaSeed: string;
  /** Simulated Soroban contract address of the requester */
  requesterAddress: string;
  status: VrfRequestStatus;
  /**
   * The generated verifiable random output (hex)
   * @nullable
   */
  randomOutput: string | null;
  /** Simulated VRF oracle contract address */
  contractAddress: string;
  /** Simulated gas cost for on-chain verification */
  gasEstimate: number;
  createdAt: string;
  /** @nullable */
  fulfilledAt: string | null;
}

export type VrfRequestDetailStatus =
  (typeof VrfRequestDetailStatus)[keyof typeof VrfRequestDetailStatus];

export const VrfRequestDetailStatus = {
  pending: "pending",
  fulfilled: "fulfilled",
  failed: "failed",
} as const;

export type VrfProofVerificationStatus =
  (typeof VrfProofVerificationStatus)[keyof typeof VrfProofVerificationStatus];

export const VrfProofVerificationStatus = {
  unverified: "unverified",
  verified: "verified",
  invalid: "invalid",
} as const;

export interface VrfProof {
  id: number;
  requestId: number;
  /** EC point gamma (part of the ECVRF proof) */
  gammaPoint: string;
  /** Challenge scalar c in the proof */
  challengeScalar: string;
  /** Response scalar s in the proof */
  responseScalar: string;
  /** VRF public key (hex) */
  publicKey: string;
  /** Full serialized proof (hex) */
  proofBytes: string;
  verificationStatus: VrfProofVerificationStatus;
  /**
   * JSON array of verification steps
   * @nullable
   */
  verificationSteps: string | null;
  computedAt: string;
}

export interface VrfRequestDetail {
  id: number;
  alphaSeed: string;
  requesterAddress: string;
  status: VrfRequestDetailStatus;
  /** @nullable */
  randomOutput: string | null;
  contractAddress: string;
  gasEstimate: number;
  createdAt: string;
  /** @nullable */
  fulfilledAt: string | null;
  proof?: VrfProof;
}

export interface CreateVrfRequestBody {
  /** Input seed for randomness generation */
  alphaSeed: string;
  /** Soroban contract address of the requester */
  requesterAddress: string;
}

export interface VerificationStep {
  stepNumber: number;
  name: string;
  description: string;
  passed: boolean;
  detail: string;
}

export interface VerificationResult {
  proofId: number;
  valid: boolean;
  steps: VerificationStep[];
  gasUsed: number;
  /** Simulated block time in milliseconds */
  blockTime: number;
}

export interface DashboardStats {
  totalRequests: number;
  pendingRequests: number;
  fulfilledRequests: number;
  failedRequests: number;
  /** Percentage of proofs that passed verification */
  proofSuccessRate: number;
  avgGasPerVerification: number;
  avgFulfillmentTimeMs: number;
  totalProofsGenerated: number;
  totalProofsVerified: number;
}

export interface RandomnessDistributionPoint {
  /** Hex range bucket */
  bucket: string;
  count: number;
  percentage: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  request_created: "request_created",
  proof_generated: "proof_generated",
  proof_verified: "proof_verified",
  request_fulfilled: "request_fulfilled",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  requestId: number | null;
  /** @nullable */
  proofId: number | null;
  timestamp: string;
}

export type ListVrfRequestsParams = {
  status?: ListVrfRequestsStatus;
  limit?: number;
};

export type ListVrfRequestsStatus =
  (typeof ListVrfRequestsStatus)[keyof typeof ListVrfRequestsStatus];

export const ListVrfRequestsStatus = {
  pending: "pending",
  fulfilled: "fulfilled",
  failed: "failed",
} as const;