# lib/api-zod/src/generated/types/createVrfRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Soroban VRF API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateVrfRequestBody {
  /** Input seed for randomness generation */
  alphaSeed: string;
  /** Soroban contract address of the requester */
  requesterAddress: string;
}