# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Soroban VRF API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalRequests: number;
  pendingRequests: number;
  fulfilledRequests: number;
  failedRequests: number;
  /** Percentage of proofs that passed verification */
  proofSuccessRate: number;
  avgGasPerVerification: number;
  avgFulfillmentTimeMs: number;
  totalProofsGenerated: number;
  totalProofsVerified: number;
}