# lib/api-zod/src/generated/types/randomnessDistributionPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Soroban VRF API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RandomnessDistributionPoint {
  /** Hex range bucket */
  bucket: string;
  count: number;
  percentage: number;
}