# lib/api-zod/src/generated/types/verificationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Soroban VRF API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VerificationStep } from "./verificationStep";

export interface VerificationResult {
  proofId: number;
  valid: boolean;
  steps: VerificationStep[];
  gasUsed: number;
  /** Simulated block time in milliseconds */
  blockTime: number;
}