# lib/api-zod/src/generated/types/verificationStep.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Soroban VRF API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VerificationStep {
  stepNumber: number;
  name: string;
  description: string;
  passed: boolean;
  detail: string;
}