# lib/api-zod/src/generated/types/vrfProof.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Soroban VRF API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VrfProofVerificationStatus } from "./vrfProofVerificationStatus";

export interface VrfProof {
  id: number;
  requestId: number;
  /** EC point gamma (part of the ECVRF proof) */
  gammaPoint: string;
  /** Challenge scalar c in the proof */
  challengeScalar: string;
  /** Response scalar s in the proof */
  responseScalar: string;
  /** VRF public key (hex) */
  publicKey: string;
  /** Full serialized proof (hex) */
  proofBytes: string;
  verificationStatus: VrfProofVerificationStatus;
  /**
   * JSON array of verification steps
   * @nullable
   */
  verificationSteps: string | null;
  computedAt: Date;
}