# lib/api-zod/src/generated/types/vrfRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Soroban VRF API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VrfRequestStatus } from "./vrfRequestStatus";

export interface VrfRequest {
  id: number;
  /** The input seed provided by the requester */
  alphaSeed: string;
  /** Simulated Soroban contract address of the requester */
  requesterAddress: string;
  status: VrfRequestStatus;
  /**
   * The generated verifiable random output (hex)
   * @nullable
   */
  randomOutput: string | null;
  /** Simulated VRF oracle contract address */
  contractAddress: string;
  /** Simulated gas cost for on-chain verification */
  gasEstimate: number;
  createdAt: Date;
  /** @nullable */
  fulfilledAt: Date | null;
}