# lib/api-zod/src/generated/types/vrfRequestDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Soroban VRF API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VrfProof } from "./vrfProof";
import type { VrfRequestDetailStatus } from "./vrfRequestDetailStatus";

export interface VrfRequestDetail {
  id: number;
  alphaSeed: string;
  requesterAddress: string;
  status: VrfRequestDetailStatus;
  /** @nullable */
  randomOutput: string | null;
  contractAddress: string;
  gasEstimate: number;
  createdAt: Date;
  /** @nullable */
  fulfilledAt: Date | null;
  proof?: VrfProof;
}